pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "heterosense-fl"
version = "1.0.0"
description = "A multimodal simulation testbed for modality-heterogeneous federated learning research"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Xun Shao",       email = "x-shao@ieee.org" },
    { name = "Kohsuke Yamakawa" },
    { name = "Aoba Otani" },
]
keywords = [
    "federated learning", "modality heterogeneity", "simulation",
    "point cloud", "pressure map", "temporal window", "activity recognition",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov>=4.0"]
docs = ["sphinx>=7.0", "sphinx-rtd-theme>=2.0"]

[project.urls]
Homepage     = "https://github.com/ss-lab-tut/heterosense-fl"
Repository   = "https://github.com/ss-lab-tut/heterosense-fl"
Issues       = "https://github.com/ss-lab-tut/heterosense-fl/issues"
Paper        = "https://www.sciencedirect.com/journal/softwarex"

[project.scripts]
heterosense-demo      = "heterosense._scripts.quick_demo:main"
heterosense-benchmark = "heterosense._scripts.run_benchmark:main"

[tool.setuptools.packages.find]
where     = ["."]
include   = ["heterosense*"]
exclude   = ["tests*", "docs*", "examples*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"
